Add features/ dir. Move components into features
