+++ features/auth/components/Elements/AuthForm.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import Form, { FormProps } from '../ReactHookForm/Form';
+import Form, { FormProps } from '../../../../components/Elements/ReactHookForm/Form';
 
 const StyledForm = styled(Form)`
   width: 100%;
