+++ pages/signup.tsx
@@ -1,9 +1,7 @@
 // components
 import TextInput from '../components/Elements/ReactHookForm/TextInput';
-import AuthFormLayout from '../components/Layouts/AuthFormLayout';
-import AuthFormContentsWrapper from '../components/Elements/AuthForm/AuthFormContentsWrapper';
-import AuthForm from '../components/Elements/AuthForm/AuthForm';
 import Button from '../components/Elements/Button/Button';
+import { AuthForm, AuthFormContentsWrapper, AuthFormLayout } from '../features/auth/index';
 
 // validations
 import { emailRegExp } from '../const/validation/rules/email';
@@ -26,10 +24,13 @@ type SignUpFormValues = {
 
 const SignUp = () => {
   // TODO: If the user logged in, redirect to main page
+  // TODO: do this in the layout component
 
   const onSubmit = () => {
     /* do something */
     // TODO: POST: login API
+
+    // TODO: axiosBase, service, reactQueryCustomHook, API endpoint
   };
 
   // TODO: Set the same password validation as the backend
